Refactor join/leave API to allow joining before having an IP address
We may start long before we've acquired a DHCP lease, or the system has
been properly configured.  Yet we still want to be able to keep track of
any queriers on the LAN.

This patch changes the k_join() and k_leave() APIs to use the Linux (and
FreeBSD) specific `struct ip_mreqn` which takes an ifindex parameter.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>