Refactor, drop uvifs[] array, rely on interfaces from .conf file
In mrouted, which this code is forked from, we must be able to run from
all interfaces, as long as they have an IP address.  It's requirements
are the reversed of querierd.

This patch flips the logic around, we drop the uvifs[] array, rely on
the interfaces from the .conf file (user knows best), and populate all
interfaces with info from kernel: address, up/down, etc.  The resulting
code is quite a bit cleaner already, but future commits will clean it up
even further.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>