Add support for interfaces being removed/added at runtime
When starting up, all interfaces may not yet have been created in the
system.  We know the intent from the .conf file, so we can prepare for
when the interface(s) appear.  Similarily, when interfaces are removed
we drop all resources connected to the interface and mark it as DOWN.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>