Add IPC support for showing multicast groups from bridge mdb
This patch adds support for two different modes of showing the multicast
overview.  There are two sources of information for this, the bridge and
querierd itself.  The `bridge mdb show` output provides the muliticast
groups, per VLAN, and querierctl provides Querier IP address per VLAN
and their respecitve timers.

Two modes of output are implemented:

  1. `querierctl show [detail]`, which is the default (native) format
  2. `querierctl [-p ]show compat [detail]`, which is the WeOS format

Great care has been taked to ensure the latter outputs its data in a
format that the automated test system (Fawlty) at Westermo knows how
to parse -- we don't want to update all tests for this new world.

The native format is preferred by the author and should possibly be
the default output also in WeOS, but make sure to sue the compat one
in non-interactive mode, which Fawlty runs the CLI in.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>