querierctl: Read router ports from bridge-vlan command output
/sys/class/net/<BRIDGEIF>/bridge/multicast_router_ports is unlikly to
get accepted as an API for reading router ports. However the information
is available to user-space through netlink. The vlan-global
sub command of bridge(8) exposes this and is fairly easy to parse.

This commit adds bitstring.h from OpenBSD, which seems like a fair set
of bitmap macros, with the same (BSD-3-Clause) license as src/queue.h.

Signed-off-by: Jacques de Laval <Jacques.De.Laval@westermo.com>