Send proxy queries in proxy mode
This patch adds support for sending proxy queries from source 0.0.0.0.

To get around the kernel filling in an "appropriate" src address we need
to use a raw packet socket and construct the ethernet frame + a correct
IPv4 header. A separate socket, and data buffer are added for this but
most of the data in the buffer can stay static after initialization. Only
the src MAC address changes depending on which interface should send the
proxy query.

Signed-off-by: Jacques de Laval <Jacques.De.Laval@westermo.com>