querierctl: Fix handling of bigger interface indexes in router port parsing
Follow up to g97a532a

Keeping track of every possible interface index using a bitmap would
require many many bits on the stack. A better data structure would
possibly be a hashmap, but a pragmatic aproach at least for now is to
just sort the input before parsing, and since we are already fork a shell
we can just pipe to sort(1).

This patch introduces jq as a dependency. JSON is the only supported
machine friendly output format from iproute2 and jq is already available
on the current known targets of querierd.

Signed-off-by: Jacques de Laval <Jacques.De.Laval@westermo.com>