pev: Add callback private data destructor callback
The private callback data is only available is only exposed in the API
when a callback is called. In a scenario where the private data has been
dynamically allocated on the heap and the event is deleted outside of the
callback this is problematic since the user of the API has no way of
deallocating this data.

This patch adds a new optional callback that gets called during the
deletion of an event. It is then possible to do what ever clean up is
necessary in this callback.

Signed-off-by: Jacques de Laval <Jacques.De.Laval@westermo.com>